requirements.txt
asgiref==3.5.2
Brotli==1.0.9
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.2
whitenoise==6.2.0
